Run a task on ECS Github Action
Summary:
Derived from the amazon-ecs-deploy-task-definition, this seeks to
address the missing option to deploy a single task. To reduce
duplication of code and the need for maintence, I have also opted to run
this with a task-arn instead of a file. You can still use the amazon
action to only create the task in ecs, then use the output in this task.